hooks: use unsigned instead of size_t to count samples

+++ include/villas/hook.h
@@ -69,11 +69,11 @@ int hook_stop(struct hook *h);
 int hook_periodic(struct hook *h);
 int hook_restart(struct hook *h);
 
-int hook_read(struct hook *h, struct sample *smps[], size_t *cnt);
-int hook_write(struct hook *h, struct sample *smps[], size_t *cnt);
+int hook_read(struct hook *h, struct sample *smps[], unsigned *cnt);
+int hook_write(struct hook *h, struct sample *smps[], unsigned *cnt);
 
-size_t hook_read_list(struct list *hs, struct sample *smps[], size_t cnt);
-size_t hook_write_list(struct list *hs, struct sample *smps[], size_t cnt);
+int hook_read_list(struct list *hs, struct sample *smps[], unsigned cnt);
+int hook_write_list(struct list *hs, struct sample *smps[], unsigned cnt);
 
 /** Compare two hook functions with their priority. Used by list_sort() */
 int hook_cmp_priority(const void *a, const void *b);
