+++ lib/hooks/shift_ts.c
@@ -78,7 +78,7 @@ static int shift_ts_parse(struct hook *h, json_t *cfg)
 	return 0;
 }
 
-static int shift_ts_read(struct hook *h, struct sample *smps[], size_t *cnt)
+static int shift_ts_read(struct hook *h, struct sample *smps[], unsigned *cnt)
 {
 	struct shift_ts *p = h->_vd;
